Separate out views into their own files.
